+++ projects/design-angular-kit/src/lib/design-angular-kit.module.ts
@@ -3,6 +3,8 @@ import { CommonModule } from '@angular/common';
 import { FormsModule } from '@angular/forms';
 
 import { CheckboxComponent } from './checkbox/checkbox.component';
+import { PopoverDirective } from './popover/popover.directive';
+import { PopoverComponent } from './popover/popover.component';
 import { ProgressBarComponent } from './progress-bar/progress-bar.component';
 import { RadioButtonComponent, RadioGroupDirective } from './radio/radio.component';
 import { ToggleComponent } from './toggle/toggle.component';
@@ -13,6 +15,14 @@ import { BadgeDirective } from './badge/badge.directive';
 import { CollapseDirective } from './collapse/collapse.directive';
 import { CollapseItemComponent } from './collapse/collapse-item.component';
 import { CollapseGroupComponent } from './collapse/collapse-group.component';
+import { DropdownComponent } from './dropdown/dropdown.component';
+import { DropdownItemComponent } from './dropdown/dropdown-item.component';
+import { DropdownHeaderComponent } from './dropdown/dropdown-header.component';
+import { DropdownDividerComponent } from './dropdown/dropdown-divider.component';
+import { BreadcrumbComponent } from './breadcrumb/breadcrumb.component';
+import { BreadcrumbItemComponent } from './breadcrumb/breadcrumb-item.component';
+import { TabGroupComponent } from './tabs/tab-group.component';
+import { TabComponent } from './tabs/tab.component';
 
 @NgModule({
   imports: [
@@ -25,13 +35,23 @@ import { CollapseGroupComponent } from './collapse/collapse-group.component';
     RadioGroupDirective,
     RadioButtonComponent,
     BadgeDirective,
+    TabGroupComponent,
+    TabComponent,
     ProgressBarComponent,
     ButtonComponent,
     TooltipDirective,
     TooltipComponent,
     CollapseDirective,
     CollapseItemComponent,
-    CollapseGroupComponent
+    CollapseGroupComponent,
+    PopoverDirective,
+    PopoverComponent,
+    DropdownComponent,
+    DropdownItemComponent,
+    DropdownHeaderComponent,
+    DropdownDividerComponent,
+    BreadcrumbComponent,
+    BreadcrumbItemComponent
   ],
   exports: [
     CheckboxComponent,
@@ -39,13 +59,22 @@ import { CollapseGroupComponent } from './collapse/collapse-group.component';
     RadioGroupDirective,
     RadioButtonComponent,
     BadgeDirective,
+    TabGroupComponent,
+    TabComponent,
     ProgressBarComponent,
     ButtonComponent,
     TooltipDirective,
     CollapseDirective,
     CollapseItemComponent,
-    CollapseGroupComponent
+    CollapseGroupComponent,
+    PopoverDirective,
+    DropdownComponent,
+    DropdownItemComponent,
+    DropdownHeaderComponent,
+    DropdownDividerComponent,
+    BreadcrumbComponent,
+    BreadcrumbItemComponent
   ],
-  entryComponents: [TooltipComponent]
+  entryComponents: [TooltipComponent, PopoverComponent]
 })
 export class DesignAngularKitModule { }