+++ projects/design-angular-kit/src/public_api.ts
@@ -4,6 +4,8 @@
 
 export * from './lib/progress-bar/progress-bar.component';
 export * from './lib/checkbox/checkbox.component';
+export * from './lib/popover/popover.directive';
+export * from './lib/popover/popover.config';
 export * from './lib/toggle/toggle.component';
 export * from './lib/radio/radio.component';
 export * from './lib/tooltip/tooltip.directive';
@@ -13,4 +15,11 @@ export * from './lib/badge/badge.directive';
 export * from './lib/collapse/collapse.directive';
 export * from './lib/collapse/collapse-item.component';
 export * from './lib/collapse/collapse-group.component';
+export * from './lib/dropdown/dropdown.component';
+export * from './lib/dropdown/dropdown-item.component';
+export * from './lib/dropdown/dropdown-header.component';
+export * from './lib/dropdown/dropdown-divider.component';
+export * from './lib/breadcrumb/breadcrumb.component';
+export * from './lib/breadcrumb/breadcrumb-item.component';
+export * from './lib/tabs/tab-group.component';
 export * from './lib/design-angular-kit.module';