+++ src/app/popover/popover-disabled-example/popover-disabled-example.component.ts
@@ -1,0 +1,10 @@
+import { Component, OnInit } from '@angular/core';
+
+@Component({
+  selector: 'it-popover-disabled-example',
+  templateUrl: './popover-disabled-example.component.html',
+  styleUrls: ['./popover-disabled-example.component.scss']
+})
+export class PopoverDisabledExampleComponent {
+
+}